Break out editor under actions
